Update Score and Lives
Increases the score by 1 whenever a new brick is created as a result
of hitting a brick, and by 10 whenever a row is cleared. Also
decrements the number of lives by 1 when the bottom row is hit by
the ball, and resets the game when it goes below 0. These values are
all customizable.