Added Start Menu Splashscreen
A splashscreen is now shown when starting up the game (through
the Scene called MainMenu). Pressing any key will start the game.
If the player runs out of lives he will be returned to the start
menu.

Also fixes a bug in determining where to spawn bricks when hitting
the top wall.